Cast enum's to size_t's for hashing.

+++ include/boost/multiprecision/cpp_dec_float.hpp
@@ -258,7 +258,7 @@ class cpp_dec_float
       std::size_t result = 0;
       for (int i = 0; i < prec_elem; ++i)
          boost::multiprecision::detail::hash_combine(result, data[i]);
-      boost::multiprecision::detail::hash_combine(result, exp, neg, fpclass);
+      boost::multiprecision::detail::hash_combine(result, exp, neg, static_cast<std::size_t>(fpclass));
       return result;
    }
